Portable Helper Tool.
This was done to help a customer. In Sencore, we used a bash script to gather and log technical data in one of our units. To do this in a Windows environment customers have to install Cygwin and run the bash script. Some customers find it too difficult to execute. Therefore, I build this small helper GUI to make it easy.